Created new music player with music sourcing from link to Google Cloud bucket.
