+++ src/components/SoundcloudPlayer/MusicPlayer.css
@@ -1,0 +1,3 @@
+.loadTrackBtn {
+  background-color: #F2D88F;
+}
